Create a views module and separate views in multiple files
- Improve the views imports on urls.py
