+++ contacts/views/__init__.py
@@ -1,0 +1,3 @@
+from .contacts import *
+from .emails import *
+from .phone_numbers import *
